sale_triple_discount: Consolidate discount in std field
Until now, the triple discount feature did use the standard Odoo field
as the first discount field, adding only extra fields for the second
and the third discount. This implied we had to override any function
using discount to consider the other discounts properly.

By adding an extra field discount1 to store the first discount, it allows
to redefine the standard discount field to a computed field that will
consolidate the triple discount from the other fields, and avoid the
need to redefine anything relying on the discount field as it will
already consider the triple discounts.

Apply rounding on discount

Fix computed fields dependencies

The issue with actual implementation is that discount is already
a computed field in std Odoo (eg to apply discount from pricelist).

Therefore, the value computed by Odoo must be stored in another field,
(ie discount1) for discount to be used to consolidate the multiple
discount. For this, we assign _compute_discount as the compute
of discount1 and redefine a new function to compute discount as
consolidated from the other discount fields.

Manually bump major version to play upgrade script

Fix migration script import

Fix init discount from pricelist

Fix invoicing

Add error in case of dependency module version mismatch